operation area of effect estimation + refactoring
